Use BackwardLightSampler instead of LightSampler, extract LightSample from LightSampler
