+++ src/appleseed/renderer/kernel/lighting/backwardlightsampler.h
@@ -33,7 +33,8 @@
 
 // appleseed.renderer headers.
 #include "renderer/kernel/intersection/intersectionsettings.h"
-#include "renderer/kernel/lighting/lightsampler.h"
+#include "renderer/kernel/lighting/lightsample.h"
+#include "renderer/kernel/lighting/lighttree.h"
 #include "renderer/kernel/lighting/lighttypes.h"
 #include "renderer/kernel/shading/shadingray.h"
 #include "renderer/modeling/scene/containers.h"
@@ -119,6 +120,12 @@ class BackwardLightSampler
         const foundation::Vector3f&         s,
         LightSample&                        light_sample) const;
 
+    // Sample the sets of non-physical lights and emitting triangles.
+    void sample(
+        const ShadingRay::Time&             time,
+        const foundation::Vector3f&         s,
+        LightSample&                        light_sample) const;
+    
     // Sample the sets of non-physical lights and emitting triangles using a light-tree.
     void sample(
         const ShadingRay::Time&             time,
@@ -232,7 +239,7 @@ inline size_t BackwardLightSampler::get_light_tree_light_count() const
 
 inline bool BackwardLightSampler::has_lights_or_emitting_triangles() const
 {
-    return m_non_physical_lights_cdf.valid() || m_emitting_triangles_cdf.valid();
+    return m_non_physical_lights_cdf.valid() || m_emitting_triangles_cdf.valid() || m_light_tree.is_built();
 }
 
 inline void BackwardLightSampler::sample_non_physical_light(
