+++ src/appleseed/renderer/kernel/lighting/directlightingintegrator.h
@@ -46,7 +46,7 @@
 
 // Forward declarations.
 namespace renderer  { class LightSample; }
-namespace renderer  { class LightSampler; }
+namespace renderer  { class BackwardLightSampler; }
 namespace renderer  { class ShadingComponents; }
 namespace renderer  { class ShadingContext; }
 namespace renderer  { class ShadingPoint; }
@@ -75,7 +75,7 @@ class DirectLightingIntegrator
     // Constructor.
     DirectLightingIntegrator(
         const ShadingContext&       shading_context,
-        const LightSampler&         light_sampler,
+        const BackwardLightSampler&         light_sampler,
         const ShadingPoint&         shading_point,
         const IMaterialSampler&     material_sampler,
         const ShadingRay::Time&     time,
@@ -117,7 +117,7 @@ class DirectLightingIntegrator
   private:
     const ShadingContext&               m_shading_context;
     const ShadingPoint&                 m_shading_point;
-    const LightSampler&                 m_light_sampler;
+    const BackwardLightSampler&                 m_light_sampler;
     const ShadingRay::Time&             m_time;
     const IMaterialSampler&             m_material_sampler;
     const int                           m_light_sampling_modes;
