+++ src/appleseed/renderer/kernel/lighting/lighttracing/lighttracingsamplegenerator.cpp
@@ -34,7 +34,7 @@
 #include "renderer/global/globallogger.h"
 #include "renderer/global/globaltypes.h"
 #include "renderer/kernel/intersection/intersector.h"
-#include "renderer/kernel/lighting/lightsampler.h"
+#include "renderer/kernel/lighting/backwardlightsampler.h"
 #include "renderer/kernel/lighting/pathtracer.h"
 #include "renderer/kernel/lighting/pathvertex.h"
 #include "renderer/kernel/lighting/scatteringmode.h"
@@ -166,7 +166,7 @@ namespace
             const Frame&                frame,
             const TraceContext&         trace_context,
             TextureStore&               texture_store,
-            const LightSampler&         light_sampler,
+            const BackwardLightSampler&         light_sampler,
             const size_t                generator_index,
             const size_t                generator_count,
             OIIOTextureSystem&          oiio_texture_system,
@@ -511,7 +511,7 @@ namespace
         double                          m_safe_scene_diameter;  // world space
         float                           m_disk_point_prob;
 
-        const LightSampler&             m_light_sampler;
+        const BackwardLightSampler&             m_light_sampler;
         TextureCache                    m_texture_cache;
         Intersector                     m_intersector;
         Arena                           m_arena;
@@ -876,7 +876,7 @@ LightTracingSampleGeneratorFactory::LightTracingSampleGeneratorFactory(
     const Frame&            frame,
     const TraceContext&     trace_context,
     TextureStore&           texture_store,
-    const LightSampler&     light_sampler,
+    const BackwardLightSampler&     light_sampler,
     OIIOTextureSystem&      oiio_texture_system,
     OSLShadingSystem&       shading_system,
     const ParamArray&       params)
