+++ src/appleseed/renderer/kernel/lighting/pt/ptlightingengine.cpp
@@ -67,7 +67,7 @@
 #include <string>
 
 // Forward declarations.
-namespace renderer  { class LightSampler; }
+namespace renderer  { class BackwardLightSampler; }
 namespace renderer  { class PixelContext; }
 namespace renderer  { class TextureCache; }
 
@@ -195,7 +195,7 @@ namespace
         };
 
         PTLightingEngine(
-            const LightSampler&     light_sampler,
+            const BackwardLightSampler&     light_sampler,
             const ParamArray&       params)
           : m_params(params)
           , m_light_sampler(light_sampler)
@@ -290,7 +290,7 @@ namespace
 
       private:
         const Parameters                m_params;
-        const LightSampler&             m_light_sampler;
+        const BackwardLightSampler&             m_light_sampler;
 
         uint64                          m_path_count;
         Population<uint64>              m_path_length;
@@ -305,7 +305,7 @@ namespace
         struct PathVisitorBase
         {
             const Parameters&           m_params;
-            const LightSampler&         m_light_sampler;
+            const BackwardLightSampler&         m_light_sampler;
             SamplingContext&            m_sampling_context;
             const ShadingContext&       m_shading_context;
             const EnvironmentEDF*       m_env_edf;
@@ -314,7 +314,7 @@ namespace
 
             PathVisitorBase(
                 const Parameters&       params,
-                const LightSampler&     light_sampler,
+                const BackwardLightSampler&     light_sampler,
                 SamplingContext&        sampling_context,
                 const ShadingContext&   shading_context,
                 const Scene&            scene,
@@ -361,7 +361,7 @@ namespace
         {
             PathVisitorSimple(
                 const Parameters&       params,
-                const LightSampler&     light_sampler,
+                const BackwardLightSampler&     light_sampler,
                 SamplingContext&        sampling_context,
                 const ShadingContext&   shading_context,
                 const Scene&            scene,
@@ -451,7 +451,7 @@ namespace
 
             PathVisitorNextEventEstimation(
                 const Parameters&       params,
-                const LightSampler&     light_sampler,
+                const BackwardLightSampler&     light_sampler,
                 SamplingContext&        sampling_context,
                 const ShadingContext&   shading_context,
                 const Scene&            scene,
@@ -767,7 +767,7 @@ namespace
         struct VolumeVisitorDistanceSampling
         {
             const Parameters&           m_params;
-            const LightSampler&         m_light_sampler;
+            const BackwardLightSampler&         m_light_sampler;
             SamplingContext&            m_sampling_context;
             const ShadingContext&       m_shading_context;
             ShadingComponents&          m_path_radiance;
@@ -777,7 +777,7 @@ namespace
 
             VolumeVisitorDistanceSampling(
                 const Parameters&       params,
-                const LightSampler&     light_sampler,
+                const BackwardLightSampler&     light_sampler,
                 SamplingContext&        sampling_context,
                 const ShadingContext&   shading_context,
                 const Scene&            scene,
@@ -1049,7 +1049,7 @@ namespace
 //
 
 PTLightingEngineFactory::PTLightingEngineFactory(
-    const LightSampler& light_sampler,
+    const BackwardLightSampler& light_sampler,
     const ParamArray&   params)
   : m_light_sampler(light_sampler)
   , m_params(params)
