+++ src/appleseed/renderer/modeling/shadergroup/shadergroup.h
@@ -130,6 +130,8 @@ class APPLESEED_DLLSYMBOL ShaderGroup
     void* osl_shader_group() const;
 
   private:
+    friend class BackwardLightSampler;
+    // friend class ForwardLightSampler;
     friend class LightSampler;
     friend class ShaderGroupFactory;
